+++ bigdata_briefs/__main__.py
@@ -1,7 +1,10 @@
 from bigdata_briefs import LOG_LEVEL
 from bigdata_briefs.api.app import app
+from bigdata_briefs.settings import settings
 
 if __name__ == "__main__":
     import uvicorn
 
-    uvicorn.run(app, host="0.0.0.0", port=8001, log_level=LOG_LEVEL.lower())
+    uvicorn.run(
+        app, host=settings.HOST, port=settings.PORT, log_level=LOG_LEVEL.lower()
+    )