+++ bigdata_briefs/utils.py
@@ -1,5 +1,7 @@
 import random
 import time
+import traceback
+import warnings
 from datetime import datetime
 from functools import wraps
 from time import perf_counter
@@ -68,12 +70,12 @@ def validate_and_repair_model(json_str: str, model: Type[BaseModel]) -> BaseMode
         try:
             response = model.model_validate_json(fixed_json_str)
             logger.debug(
-                f"The following could not be parsed as a {model.__name__}, but we could repair the json {json_str=}"
+                f"The following could not be parsed as a {model.__name__}, but we could repair the json\n{json_str=}\n{fixed_json_str=}"
             )
             return response
         except ValidationError:
             logger.warning(
-                f"The following LLM completion could not be parsed as a {model.__name__}, nor could it be repaired"
+                f"The following LLM completion could not be parsed as a {model.__name__}, nor could it be repaired\n{json_str=}\n{fixed_json_str=}"
             )
             raise
 
@@ -95,3 +97,11 @@ def sleep_with_backoff(*, base: int = 1, attempt: int):
     logger.debug(f"Sleeping for {sleep_time}")
 
     time.sleep(sleep_time)
+
+
+def raise_warning_from(e, category=RuntimeWarning):
+    """Issue a warning derived from an exception object."""
+    tb_str = "".join(traceback.format_exception(type(e), e, e.__traceback__))
+    warnings.warn(
+        f"Converted exception to warning:\n{tb_str.strip()}", category, stacklevel=2
+    )